Shared code to execute intermediate and blockbatch models
